Bug 784912 - nsIStreamListener.onDataAvailable should handle 64-bit offset. r=standard8

+++ mailnews/base/src/nsCopyMessageStreamListener.cpp
@@ -87,7 +87,7 @@ NS_IMETHODIMP nsCopyMessageStreamListener::EndMessage(nsMsgKey key)
 }
 
 
-NS_IMETHODIMP nsCopyMessageStreamListener::OnDataAvailable(nsIRequest * /* request */, nsISupports *ctxt, nsIInputStream *aIStream, uint32_t sourceOffset, uint32_t aLength)
+NS_IMETHODIMP nsCopyMessageStreamListener::OnDataAvailable(nsIRequest * /* request */, nsISupports *ctxt, nsIInputStream *aIStream, uint64_t sourceOffset, uint32_t aLength)
 {
 	nsresult rv;
 	rv = mDestination->CopyData(aIStream, aLength);
